+++ npm/finder-lib/src/index.ts
@@ -3,7 +3,7 @@
 
 // Version which is fitting with the finder version
 
-import newest = versions.v0_3_17;
+import newest = versions.v0_3_18;
 export { newest };
 
 /// ALL Versions
@@ -240,6 +240,7 @@ export namespace versions {
       description?: string;
       tags?: string[];
       name: string;
+      name_regex?: string;
       folders?: Folder[];
       files?: File[] | string[];
       command?: string;
@@ -252,6 +253,7 @@ export namespace versions {
 
     export interface Folder {
       name: string;
+      name_regex?: string;
       folders?: Folder[];
       files?: File[] | string[];
       command?: string;
@@ -261,9 +263,66 @@ export namespace versions {
 
     export interface File {
       name: string;
+      name_regex?: string;
       existence?: Existence;
       size?: Size;
       checksums?: Checksums;
+    }
+
+    export interface Size {
+      min?: number;
+      max?: number;
+      min_size_type?: SizeType;
+      max_size_type?: SizeType;
+    }
+
+    export interface Checksums {
+      sha256?: string;
+      sha512?: string;
+    }
+
+    export type Existence = "required" | "forbidden" | "optional";
+
+    export type SizeType = "B" | "KB" | "MB" | "GB";
+  }
+
+  /// Version v0.3.19
+  export namespace v0_3_19 {
+    export interface Template {
+      min_version?: string;
+      description?: string;
+      tags?: string[];
+      name: string;
+      name_regex?: string;
+      folders?: Folder[];
+      files?: File[] | string[];
+      command?: string;
+      invert_command?: boolean;
+      size?: Size;
+      mdnote?: string;
+      author?: string;
+      authors?: string[];
+      excluded_names?: string[];
+    }
+
+    export interface Folder {
+      name: string;
+      name_regex?: string;
+      folders?: Folder[];
+      files?: File[] | string[];
+      command?: string;
+      invert_command?: boolean;
+      size?: Size;
+      excluded_names?: string[];
+    }
+
+    export interface File {
+      name: string;
+      name_regex?: string;
+      existence?: Existence;
+      size?: Size;
+      checksums?: Checksums;
+      excluded_names?: string[];
     }
 
     export interface Size {