+++ npm/site/pages/index.tsx
@@ -1,4 +1,4 @@
-import { jsx, Fragment, raw } from "../src/jsx-runtime";
+import { jsx, Fragment, raw } from "@twine/core/jsx-runtime";
 import { SERVER_ADRESS } from "../src/vars";
 
 export default function render(el: HTMLDivElement, data: string) {
@@ -20,9 +20,19 @@ export default function render(el: HTMLDivElement, data: string) {
           <li>
             <a href="./regexcreator">Regex Creator</a>
           </li>
+          {import.meta.env.MODE == "backend" && (
+            <li>
+              <a href="./console">Console</a>
+            </li>
+          )}
           <li>
             <a href="./tools">Tools</a>
           </li>
+          {import.meta.env.MODE == "backend" && (
+            <li>
+              <a href="./mcapp">MCAPP</a>
+            </li>
+          )}
           {/* {import.meta.env.MODE == "backend" && (
             <li>
               <a href="./cacheviewer">Cache Viewer</a>