+++ npm/site/pages/mcapp/worlds.jsx
@@ -1,4 +1,4 @@
-import { escapeHtml, jsx, raw, Fragment } from "../../src/jsx-runtime";
+import { escapeHtml, jsx, raw, Fragment } from "@twine/core/jsx-runtime";
 import Header from "../../src/components/mcappheader.jsx";
 import * as f from "../../src/components/mcappstyle.js";
 import { SERVER_ADRESS } from "../../src/vars.js";