+++ npm/site/src/main.ts
@@ -1,6 +1,6 @@
 import { pages } from "virtual:pages";
 import { ErrorPage, render404, render404_2, render404_3 } from "./404";
-import { loadStyles } from "./jsx-runtime";
+import { loadStyles } from "@twine/core/jsx-runtime";
 import "./base.css";
 import { renderMarkdown } from "./markdown";